In train loop printing out proportion of generated words that are also in train or assess
Now after each epoch I generate 100 words and print out the percentage of them that are in the training set, dev set, and are novel words. Once the model starts overfitting it looks like the proportion of words from the train set skyrockets immediately, but I want to actually see when it happens.