Fixing min_prob and probabilities returned by /next_probs
There was a failing test because of how the app handles min_prob. Fixing this, and returning probabilities that are truncated floats rather than ints (e.g. 1.43 instead of 1, from an original 1.4312323). This should make the chart presentation more sensible: we don't need a lot of precision, but we also don't want to equate .3 and .9.